refactor: useFetch 정의 및 타입 명시
- 반복호출되는 로직을 useFetch로 분리했습니다.
- 반복되는 상수나 타입을 정의했습니다.
