+++ src/constants/order.ts
@@ -1,3 +1,7 @@
 export const TODAY = '2023-03-08';
 export const FILTERS = ['field', 'page', 'status', 'query', 'sort'];
 export const DELAY = 5000;
+export const FIELD = {
+  today: 'today',
+  all: 'all',
+} as const;
